created separate file for class Country
