+++ Day-04/loan_eligibility.rb
@@ -1,5 +1,4 @@
 module CountryLoanEligibility
-
   def check_loan_eligibility(country_name, gdp)
     if gdp > 1000000
       "#{country_name} is Eligible for Loan"
