+++ Day-04/un_eligibility.rb
