+++ Day-04/war_fitness.rb
